+++ db/migrate/20180420035507_create_moons.rb
@@ -2,10 +2,10 @@ class CreateMoons < ActiveRecord::Migration[5.1]
   def change
     create_table :moons do |t|
       t.text :name
-      t.integer :distance_planet
-      t.integer :radius_orbit_planet
-      t.integer :radius_moon
-      t.integer :orbit_planet
+      t.float :distance_planet
+      t.float :radius_orbit_planet
+      t.float :radius_moon
+      t.float :orbit_planet
       t.text :image
       t.integer :planet_id
       t.integer :user_id
